Plus/minus buttons enabled after reset
The plus and minus button did not work after reseting the
clocks do to the fact that the playing variable was still set to
'true'.

I set the playing variable to 'false' after a clock times out so
that the plus/minus buttons can be used to change the time.